requirements.txt
appdirs==1.4.4
attrs==21.2.0
cattrs==1.8.0
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.11.0
gmplot==1.4.1
gpxpy==1.4.2
idna==3.3
kiwisolver==1.3.2
matplotlib
numpy
Pillow==8.4.0
pyparsing==3.0.4
pyproj
python-dateutil==2.8.2
requests==2.26.0
requests-cache==0.8.1
scipy
six==1.16.0
url-normalize==1.4.3
urllib3==1.26.7
